Modify fetchAirly function to return data instead of response
The 'fetchAirly' function in 'rest_client.go' has been modified to read all data from the response body, close it, and return the data as bytes. Consequently, the 'parseJsonResponse' function in 'response_parser.go' now takes a byte array as an argument instead of an HTTP response. This makes parsing responses more efficient as the data is already read.